Move trait to a more generic module
