+++ thoth-api/src/lib.rs
@@ -23,6 +23,8 @@ pub mod graphql;
 pub mod imprint;
 pub mod issue;
 pub mod language;
+#[macro_use]
+pub mod model;
 pub mod price;
 pub mod publication;
 pub mod publisher;
